Elves have a strict ordered hierarchy (idno?)
I love how the input for this day was meticulously crafted to
have [[2]] & [[6]] as SINGLE occurrences in one of the ~300
nested packets, fully taking advantage of people like me who
fail to properly read the question.

Of course in plain English it said we were to add the two
packets [[2]] & [[6]] to our other packets, for a total of 302 - but
what did I do? Didn't read and assumed we should be
searching for these packets inside our existing packets,
after all when I did a quick ctrl + f there were unique
occurrences of each!!!